Switch types around
